refactor: review typescript signature (#164)
