+++ package.json
@@ -13,7 +13,7 @@
     "lint": "standard | snazzy",
     "lint:fix": "standard --fix | snazzy",
     "pretest": "npm run clean",
-    "test": "npm run lint && npm run unit && npm run typescript",
+    "test": "npm run lint && npm run unit && npm run typescript && ts-node test/types/index.ts",
     "posttest": "npm run clean",
     "typescript": "tsd",
     "prepare": "husky install",
@@ -50,10 +50,14 @@
     "husky": "^8.0.3",
     "snazzy": "^9.0.0",
     "standard": "^17.1.0",
+    "ts-node": "^10.9.1",
     "tsd": "^0.29.0",
     "typescript": "^5.2.2"
   },
   "publishConfig": {
     "access": "public"
+  },
+  "tsd": {
+    "directory": "test/types"
   }
 }
